Sudhakar, Abishek | #Bahmni2383 | Feature to limit the number of retry attempts for a failed event
Problem:
When an event fails, it is repeated indefinitely to fix the failures because of race condition across feeds. But, for errors which are not because of race condition, it runs forever and every failed attempt increases the person_id sequence number.

Solution:
1. In this commit, a column 'retries' is added to the failed event to record the times an event is retried. If it reaches the max (as per the config), the event will not be picked again for retrying
2. Every failed attempt will record the retry failure information in the new table called 'failed_event_retry_log' table
3. If a failed event succeeds during the retry, when it is deleted from the 'failed_events' table, all its corresponding previous attempt records from the 'failed_event_retry_log' will also be deleted

NOTE: In addition to the above, we also removed the in-memory implementations of data store after discussing with Vivek